fix: collapse nested if blocks, remove redundant closure/cast/format/min_or_max in jcode-app-core
Fix all 14 clippy warnings that caused Quality Guardrails failure:
- Collapse nested if blocks into let chains (agent.rs, dcg_bridge.rs, execution_policy.rs)
- Replace redundant closure call with direct if-else (provider_control.rs)
- Replace useless format!() with .to_string() (render.rs)
- Remove unnecessary i32 cast (agentgrep.rs)
- Remove unnecessary .max(0) after saturating_sub (propose_hashline_edit.rs)

Generated-By: looper 0.0.0-dev (runner=fixer, agent=hermes)

+++ crates/jcode-app-core/src/tool/agentgrep.rs
@@ -684,7 +684,7 @@ pub fn run_smart_ffs(root: &Path, query: &SmartQuery, args: &SmartArgs) -> Resul
                 .as_ref()
                 .map(|o| o.language.clone())
                 .unwrap_or_default(),
-            score: ff.score as i32,
+            score: ff.score,
             why: vec![format!("path matched: {}", subject)],
             context_applied: None,
             structure: SmartStructure {
